+++ ansible/regtest/library/testcases/bgp/test_bgp_local_pref_weight.py
@@ -20,6 +20,8 @@
 
 import shlex
 
+import time
+
 from collections import OrderedDict
 
 from ansible.module_utils.basic import AnsibleModule
@@ -129,7 +131,7 @@ def execute_commands(module, cmd):
     """
     global HASH_DICT
 
-    if 'service' in cmd and 'restart' in cmd:
+    if 'service' in cmd and 'restart' in cmd or module.params['dry_run_mode']:
         out = None
     else:
         out = run_cli(module, cmd)
@@ -142,6 +144,52 @@ def execute_commands(module, cmd):
 
     return out
 
+def verift_bgp_pref_weight(module):
+    spine_list = module.params['spine_list']
+    switch_name = module.params['switch_name']
+    leaf_list = module.params['leaf_list']
+    global RESULT_STATUS, HASH_DICT
+    failure_summary = ''
+    cmd = "vtysh -c 'sh ip bgp'"
+    out = execute_commands(module, cmd)
+
+    if out:
+        for switch in spine_list + leaf_list:
+            network = '192.168.{}.1'.format(switch[-2::])
+            if network not in out:
+                RESULT_STATUS = False
+                failure_summary += 'On Switch {} bgp route '.format(
+                    switch_name)
+                failure_summary += 'for network {} is not present '.format(
+                    network)
+                failure_summary += 'in the output of command {}\n'.format(
+                    cmd)
+
+        local_pref = module.params['local_pref']
+        weight = module.params['weight']
+
+        if local_pref:
+            value = local_pref
+            name = 'local preference'
+        else:
+            value = weight
+            name = 'weight'
+
+        if value not in out:
+            RESULT_STATUS = False
+            failure_summary += 'On Switch {} {} '.format(switch_name, name)
+            failure_summary += 'value {} is not present '.format(value)
+            failure_summary += 'in the output of command {}\n'.format(cmd)
+    else:
+        RESULT_STATUS = False
+        failure_summary += 'On switch {} '.format(switch_name)
+        failure_summary += 'result cannot be verified since '
+        failure_summary += 'output of command {} '.format(cmd)
+        failure_summary += 'is None'
+
+    alist = [True if RESULT_STATUS else False]
+    alist.append(failure_summary)
+    return alist
 
 def verify_bgp_local_pref_weight(module):
     """
@@ -149,12 +197,11 @@ def verify_bgp_local_pref_weight(module):
     :param module: The Ansible module to fetch input parameters.
     """
     global RESULT_STATUS, HASH_DICT
-    failure_summary = ''
     switch_name = module.params['switch_name']
     pref_wt_switch = module.params['pref_wt_switch']
-    spine_list = module.params['spine_list']
-    leaf_list = module.params['leaf_list']
     package_name = module.params['package_name']
+    delay = module.params['delay']
+    retries = module.params['retries']
 
     # Get the current/running configurations
     execute_commands(module, "vtysh -c 'sh running-config'")
@@ -165,44 +212,15 @@ def verify_bgp_local_pref_weight(module):
 
     if switch_name == pref_wt_switch:
         # Get all ip bgp routes
-        cmd = "vtysh -c 'sh ip bgp'"
-        out = execute_commands(module, cmd)
-
-        if out:
-            for switch in spine_list + leaf_list:
-                network = '192.168.{}.1'.format(switch[-2::])
-                if network not in out:
-                    RESULT_STATUS = False
-                    failure_summary += 'On Switch {} bgp route '.format(
-                        switch_name)
-                    failure_summary += 'for network {} is not present '.format(
-                        network)
-                    failure_summary += 'in the output of command {}\n'.format(
-                        cmd)
-
-            local_pref = module.params['local_pref']
-            weight = module.params['weight']
-
-            if local_pref:
-                value = local_pref
-                name = 'local preference'
+        retry = retries - 1
+        while(retry):
+            if verift_bgp_pref_weight(module)[0]:
+                break
             else:
-                value = weight
-                name = 'weight'
+                time.sleep(delay)
+                retry -= 1
 
-            if value not in out:
-                RESULT_STATUS = False
-                failure_summary += 'On Switch {} {} '.format(switch_name, name)
-                failure_summary += 'value {} is not present '.format(value)
-                failure_summary += 'in the output of command {}\n'.format(cmd)
-        else:
-            RESULT_STATUS = False
-            failure_summary += 'On switch {} '.format(switch_name)
-            failure_summary += 'result cannot be verified since '
-            failure_summary += 'output of command {} '.format(cmd)
-            failure_summary += 'is None'
-
-    HASH_DICT['result.detail'] = failure_summary
+    HASH_DICT['result.detail'] = verift_bgp_pref_weight(module)[1]
 
     # Get the GOES status info
     execute_commands(module, 'goes status')
@@ -221,34 +239,57 @@ def main():
             package_name=dict(required=False, type='str'),
             hash_name=dict(required=False, type='str'),
             log_dir_path=dict(required=False, type='str'),
+            delay=dict(required=False, type='int', default=10),
+            retries=dict(required=False, type='int', default=6),
+            dry_run_mode=dict(required=False, type='bool', default=False),
         )
     )
 
     global HASH_DICT, RESULT_STATUS
+    if module.params['dry_run_mode']:
+        package_name = module.params['package_name']
+        cmds_list = []
 
-    verify_bgp_local_pref_weight(module)
+        execute_commands(module, "vtysh -c 'sh running-config'")
 
-    # Calculate the entire test result
-    HASH_DICT['result.status'] = 'Passed' if RESULT_STATUS else 'Failed'
+        # Restart and check package status
+        execute_commands(module, 'service {} restart'.format(package_name))
+        execute_commands(module, 'service {} status'.format(package_name))
+        cmd = "vtysh -c 'sh ip bgp'"
+        execute_commands(module, cmd)
+        execute_commands(module, 'goes status')
 
-    # Create a log file
-    log_file_path = module.params['log_dir_path']
-    log_file_path += '/{}.log'.format(module.params['hash_name'])
-    log_file = open(log_file_path, 'w')
-    for key, value in HASH_DICT.iteritems():
-        log_file.write(key)
-        log_file.write('\n')
-        log_file.write(str(value))
-        log_file.write('\n')
-        log_file.write('\n')
+        for key, value in HASH_DICT.iteritems():
+            cmds_list.append(key)
 
-    log_file.close()
+        # Exit the module and return the required JSON.
+        module.exit_json(
+            cmds=cmds_list
+        )
+    else:
+        verify_bgp_local_pref_weight(module)
 
-    # Exit the module and return the required JSON.
-    module.exit_json(
-        hash_dict=HASH_DICT,
-        log_file_path=log_file_path
-    )
+        # Calculate the entire test result
+        HASH_DICT['result.status'] = 'Passed' if RESULT_STATUS else 'Failed'
+
+        # Create a log file
+        log_file_path = module.params['log_dir_path']
+        log_file_path += '/{}.log'.format(module.params['hash_name'])
+        log_file = open(log_file_path, 'w')
+        for key, value in HASH_DICT.iteritems():
+            log_file.write(key)
+            log_file.write('\n')
+            log_file.write(str(value))
+            log_file.write('\n')
+            log_file.write('\n')
+
+        log_file.close()
+
+        # Exit the module and return the required JSON.
+        module.exit_json(
+            hash_dict=HASH_DICT,
+            log_file_path=log_file_path
+        )
 
 if __name__ == '__main__':
     main()
